Change int to double for complex numbers

+++ includes/ft_complex.h
@@ -6,7 +6,7 @@
 /*   By: jmaia <jmaia@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/10 10:21:07 by jmaia             #+#    #+#             */
-/*   Updated: 2021/12/10 10:34:00 by jmaia            ###   ########.fr       */
+/*   Updated: 2021/12/10 10:58:09 by jmaia            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 typedef struct s_complex
 {
-	int	x;
-	int	y;
+	double	x;
+	double	y;
 }	t_complex;
 
 t_complex	sum(t_complex, t_complex);
